refactor: #261 피드백 반영
- Integer 타입으로 제공되는 함수형 인터페이스 사용
- calculate() 메서드 파라미터 개수 축소
이 부분을 위해 따로 클래스를 만들어야 하는지 고민..
